requirements.txt
torch>=2.0.1
spacy>=3.5.4
pyvi>=0.1.1
tqdm>=4.65.0
pandas>=1.5.3
pyyaml>=6.0
matplotlib>=3.7.1
streamlit>=1.24.1